[GTK][Skia] Limit output file format to PDF https://bugs.webkit.org/show_bug.cgi?id=272913
Reviewed by Michael Catanzaro and Carlos Garcia Campos.

Skia only supports printing to PDF files, but the Print portal assumes
that apps can print to PDF, PS, and SVG formats. If the user selects
PS or SVG through the Print portal dialog, and WebKit is using Skia,
we hit an assertion failure.

The next release of XDG Desktop Portal has new API [1] in the Print
portal made specifically to aid this, we can use it for Skia builds
and if the portal version matches.

[1] https://github.com/flatpak/xdg-desktop-portal/pull/1322/

* Source/WebKit/UIProcess/API/gtk/WebKitPrintOperation.cpp:
(webkitPrintOperationPreparePrint):

Canonical link: https://commits.webkit.org/277806@main