Webpage crashes when switching from PIP to in-window https://bugs.webkit.org/show_bug.cgi?id=272943 rdar://126489417
Reviewed by Jer Noble.

When the video goes into PIP, the mode bitfield belonging to VideoPresentationInterfaceMac
Has its PIP bit set to 1. Then, when switching from PIP to in-window mode, the in-window
Bit is set to 1 before the PIP bit is cleared. During this time when both bits are turned
On, the interface attempts to switch to a fullscreen mode that does not exist. To fix this,
In VideoPresentationInterfaceMac::setMode and VideoPresentationInterfaceMac::clearMode we should
Check if any two bits are both true where one of them is the PIP bit. If this is the case,
Do not attempt to change the fullscreen mode and wait for the next instance of one of these
Functions being called.

* Source/WebCore/platform/mac/VideoPresentationInterfaceMac.mm:
(WebCore::VideoPresentationInterfaceMac::setMode):
(WebCore::VideoPresentationInterfaceMac::clearMode):

Canonical link: https://commits.webkit.org/277814@main