Avoid creating compositing scrolling layers for zero-sized scrollers https://bugs.webkit.org/show_bug.cgi?id=273110 rdar://113980927
Reviewed by NOBODY (OOPS!).

Some CMSs can create large tables with many columns which may have < 1px widths. If the cells in these columns have scrollable
overflow, then we can create composited scrollers for many layers which have zero height or width, making them not interactive.
In such cases, we can avoid the overhead of composited scrolling, which also avoids the backing sharing code which has poor
performance with many sibling scrollers.

Fix by adjusting `RenderLayerScrollableArea::canUseCompositedScrolling()` to return false if the clientWidth or clientHeight
of the layer is zero (this is the area used by scrollable content).

Because `canUseCompositedScrolling()` now depends on layout, we need to have `RenderLayer::updateLayerPosition()` set the
"NeedsPostLayoutCompositingUpdate" bit on the RenderLayer when the client size changes. The existing code keyed off of
the layer size changing for a similar invalidation relating to composited widgets, but this was incorrect, as the code
in question actually checks the content box size, so fix both code paths to use a "size at last layout" caching strategy.

* LayoutTests/compositing/scrolling/async-overflow-scrolling/no-compositing-for-zero-sized-scrollers-expected.txt: Added.
* LayoutTests/compositing/scrolling/async-overflow-scrolling/no-compositing-for-zero-sized-scrollers.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateLayerPosition):
(WebCore::RenderLayer::calculateClipRects const):
* Source/WebCore/rendering/RenderLayer.h:
* Source/WebCore/rendering/RenderLayerScrollableArea.cpp:
(WebCore::RenderLayerScrollableArea::canUseCompositedScrolling const):