[iOS] CRASH: exception thrown from AVVideoCaptureSource::stopSession() https://bugs.webkit.org/show_bug.cgi?id=273088 rdar://126717672
Reviewed by Eric Carlson.

Crashtracer data shows that an exception is thrown from -[AVCaptureSession stopRunning] because that
method is called after -beginConfiguration, but before -commitConfiguration. Make this method robust
to that exception, and add our own ASSERT that we are not in the middle of configuration in order to
catch this case before it gets to AVCapture.

* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.mm:
(WebCore::AVVideoCaptureSource::stopSession):

Canonical link: https://commits.webkit.org/277869@main