[webkitpy] Wrong simulator sometimes chosen when a device is named "Managed 0" https://bugs.webkit.org/show_bug.cgi?id=273105 rdar://126884847
Reviewed by Jonathan Bedard.

When running webkit tests in a simulator, webkitpy needs to either find or create a device
that satisfies the request. For instance, if a user passes the `--ios-simulator` flag,
webkitpy needs to either find a booted simulator that matches the request, or create and
boot such a device.

By default, devices are named "Managed X", where X is an integer that increments with each
simulator that is booted. When webkitpy was created, iOS was the only simulator, so the
only check to validate a simulator after creation is that the name matches the next
available iteration of X.

Now that there are more types of OS simulators, having, for instance, a visionOS simulator
with the name "Managed 0" will override the --ios-simulator flag such that the existing
visionOS simulator is booted instead of the newly created iOS simulator with the same name.

This PR adds extra heuristics to check the device's type and OS build version before
assuming it's a match. This locks the simulator device and OS to the requested version,
ensuring the correct simulator is always booted.

* Tools/Scripts/webkitpy/xcode/simulated_device.py:
(SimulatedDeviceManager._create_or_find_device_for_request): Added check for device type and OS build version when finding a newly created simulated device.

Canonical link: https://commits.webkit.org/277884@main