Show passed output (if available) in ScriptError. rdar://126966101 https://bugs.webkit.org/show_bug.cgi?id=273169
Reviewed by NOBODY (OOPS!).

The constructor for webkitpy.common.system.executive.ScriptError takes an `output` parameter.
This should also be displayed if a custom error message isn't present, as it often contains
more information that may be useful to the user.

* Tools/Scripts/webkitpy/common/system/executive.py:
(ScriptError.__init__): Show `output` in addition to other already shown information when no message is specified.