Regression(276851@main) GPUProcessProxy::sendBookmarkDataForCacheDirectory() hangs the main thread https://bugs.webkit.org/show_bug.cgi?id=273164 rdar://126946636
Reviewed by Per Arne Vollan.

276851@main started calling GPUProcessProxy::sendBookmarkDataForCacheDirectory() on the main thread
as soon as the GPUProcess finishes launches. The issue is that this function does file system
operations which may hang the UIProcess' main thread. This results in an app launch regression for
certain apps.

To address the issue, I moved the filesystem operation and the IPC to a background queue.

* Source/WebKit/UIProcess/Cocoa/GPUProcessProxyCocoa.mm:
(WebKit::GPUProcessProxy::sendBookmarkDataForCacheDirectory):

Canonical link: https://commits.webkit.org/277923@main