in in-window mode, video play/pauses when moving media controls https://bugs.webkit.org/show_bug.cgi?id=273127 rdar://126300898
Reviewed by Jer Noble.

On youtube.com, the click event on the media controls propagates down to the video element,
Where YouTube handles the event by toggling play/pause on the video. To fix this, we do
Not want events on the media controls to propagate down. And so the call to stopPropagation() on
The event should be moved from the button class to the macOSFullscreenMediaControls class.

* Source/WebCore/Modules/modern-media-controls/controls/button.js:
(Button.prototype.handleEvent):
* Source/WebCore/Modules/modern-media-controls/controls/macos-fullscreen-media-controls.js:
(MacOSFullscreenMediaControls.prototype.handleEvent):

Canonical link: https://commits.webkit.org/277938@main