Improve error reporting on simulated device boot failure. https://bugs.webkit.org/show_bug.cgi?id=273216 rdar://127015274
Reviewed by Jonathan Bedard.

If an error occurs while booting a simulated device, the actual error encountered
as a result of `xcrun simctl boot <UDID>` is hidden. This output is already passed
to ScriptError when it's invoked, so we should catch this exception and print this
information to the log.

* Tools/Scripts/webkitpy/xcode/simulated_device.py:
(SimulatedDeviceManager._boot_device): Add try/except block to handle ScriptError as a result of simulated device boot failing.

Canonical link: https://commits.webkit.org/277952@main