[UnifiedPDF] Snapshots after the first page are partially dark-shaded https://bugs.webkit.org/show_bug.cgi?id=273213 rdar://126817702
Reviewed by Abrar Rahman Protyasha and Simon Fraser.

277273@main attempted to address this issue by making adding in the
white background as a fill as part of the async rendering codepath. It
turns out that this was quite not enough and snapshots of pages after
the first page are still appearing as dark. It looks like this is due
to the fact that we modify the graphics context by making it so that the
origin is at the bottom of the page before we pass it off to PDFKit.
So, instead we will start adding the fill before we perform this
coordinate space transformation.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::paintPDFContent):

Canonical link: https://commits.webkit.org/277965@main