[WPE][GTK] Improve error checking in AccessibilityAtspi https://bugs.webkit.org/show_bug.cgi?id=273247
Reviewed by NOBODY (OOPS!).

Print a warning if we fail to register an accessibility object.

We could also print warnings if we fail to emit a signal, but there are
a bunch of these cases and I don't want to fix them all, and failures
there are probably much less likely.

* Source/WebCore/accessibility/atspi/AccessibilityAtspi.cpp:
(WebCore::AccessibilityAtspi::registerRoot):
(WebCore::AccessibilityAtspi::registerObject):
(WebCore::AccessibilityAtspi::registerHyperlink):
(WebCore::AccessibilityAtspi::ensureCache):