AX: AXObjectCache::m_deferredUnconnectedNodeList needs to be clear after adding the objects to the isolated tree. https://bugs.webkit.org/show_bug.cgi?id=273296 <rdar://problem/127087786>
Reviewed by Tyler Wilcock.

This collection was not being cleared, so it would keep growing for the life time of the cache object.
Also renamed it to m_deferredUnconnectedObjects to better reflect that it is a collection of AX objects and not a list of Nodes.

* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::deferAddUnconnectedNode):
(WebCore::AXObjectCache::performDeferredCacheUpdate):
* Source/WebCore/accessibility/AXObjectCache.h:

Canonical link: https://commits.webkit.org/278037@main