Disable OffscreenCanvasRenderingContext2D's commit() method by default https://bugs.webkit.org/show_bug.cgi?id=272591 rdar://126758254
Reviewed by Matt Woodrow.

This puts OffscreenCanvasRenderingContext2D's commit() behind a runtime
setting marked testable (OffscreenCanvasDeprecatedCommitEnabled). This
ensures this method continues to be tested for the time being and can
be re-enabled should there be a need.

The goal is to remove this method and underlying implementation
completely as discussed here:
https://github.com/whatwg/html/pull/9979

* LayoutTests/inspector/canvas/recording-offscreen-canvas-2d-full-expected.txt:
* LayoutTests/inspector/canvas/resources/recording-2d.js:
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/html/canvas/OffscreenCanvasRenderingContext2D.idl:

Canonical link: https://commits.webkit.org/278047@main