[visionOS] After returning to element fullscreen from LMK fullscreen, unable to re-enter LMK fullscreen https://bugs.webkit.org/show_bug.cgi?id=273344 rdar://127137182
Reviewed by Eric Carlson.

When exiting LinearMediaKit fullscreen the VideoPresentationInterfaceLMK and its
LMPlayableViewController are destroyed, but WKFullScreenViewController still has a reference to a
UIView created by the LMPlayableViewController. Since the LMPlayableViewController and its
Playable-conforming delegate no longer exist, the view has no delegate to notify when the user
interacts with it. This results in an inability to re-enter LinearMediaKit fullscreen after
previously exiting from it to element fullscreen.

Resolved this by plumbing a "didCleanupFullscreen" call from VideoPresentationInterfaceIOS to
WKFullScreenViewController. When this method is called and the VC is still in fullscreen mode,
WKFullScreenViewController is told to remove the old _environmentPickerButtonView and create a new
one (which will re-ensure a VideoPresentationInterfaceLMK and LMPlayableViewController).

* Source/WebKit/UIProcess/Cocoa/VideoPresentationManagerProxy.mm:
(WebKit::VideoPresentationManagerProxy::didCleanupFullscreen):
* Source/WebKit/UIProcess/PageClient.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didCleanupFullscreen):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/ios/PageClientImplIOS.h:
* Source/WebKit/UIProcess/ios/PageClientImplIOS.mm:
(WebKit::PageClientImpl::didCleanupFullscreen):
* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenViewController.h:
* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenViewController.mm:
(-[WKFullScreenViewController videoControlsManagerDidChange]):
(-[WKFullScreenViewController configureEnvironmentPickerButtonView]):
(-[WKFullScreenViewController _configureEnvironmentPickerButtonViewWithPlayableViewController:]): Deleted.
* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.h:
* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController didCleanupFullscreen]):
* Source/WebKit/UIProcess/mac/PageClientImplMac.h:

Canonical link: https://commits.webkit.org/278089@main