WPT dom/events/shadow-relatedTarget.html fails partially https://bugs.webkit.org/show_bug.cgi?id=272300
Reviewed by Ryosuke Niwa.

This patch changes to see whether `EventContext` is window context or not
in `EventPath::setRelatedTarget`.

Probably, it might make related path simple
if EventPath have an another flag to indicate the context is for
window separately to coexit the `EventContext::Type::MouseOrFocus`,
but this patch choose only to add the condition to
it by that `EventPath::setRelatedTarget()` is called only for MouseEvent or
FocusEvent, and will be skipped essentially for other events.

* LayoutTests/imported/w3c/web-platform-tests/dom/events/shadow-relatedTarget-expected.txt:
* Source/WebCore/dom/EventContext.cpp:
(WebCore::EventContext::handleLocalEvents const):
* Source/WebCore/dom/EventContext.h:
(WebCore::EventContext::isNormalEventContext const):
* Source/WebCore/dom/EventPath.cpp:
(WebCore::EventPath::setRelatedTarget):

Canonical link: https://commits.webkit.org/278090@main