[iOS] Rebaseline several pasteboard-related layout tests https://bugs.webkit.org/show_bug.cgi?id=273378 rdar://124417777
Reviewed by Richard Robinson.

Rebaseline several layout tests in `editing/async-clipboard` and `editing/pasteboard` after system
changes (in UIKit). In iOS 17.4+, UIKit now automatically publishes `public.utf8-plain-text` on the
pasteboard as long as `public.html` is written.

The value of the plain text type is either a value explicitly written to the pasteboard, or (if no
value was specified) the text content of a document derived from loading the HTML data (by using
UIFoundation). As such, we make these tests pass in both cases where the system does (or doesn't)
automatically write plain text on behalf of the pasteboard client when HTML data is written.

* LayoutTests/editing/async-clipboard/clipboard-write-basic-expected.txt:
* LayoutTests/editing/async-clipboard/clipboard-write-basic.html:
* LayoutTests/editing/async-clipboard/sanitize-when-reading-markup-expected.txt:
* LayoutTests/editing/async-clipboard/sanitize-when-reading-markup.html:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitize-url-when-copying-in-null-origin-expected.txt:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitize-url-when-copying-in-null-origin.html:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitizes-html-when-copying-expected.txt:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitizes-html-when-copying-in-null-origin-expected.txt:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitizes-html-when-copying-in-null-origin.html:
* LayoutTests/editing/pasteboard/data-transfer-set-data-sanitizes-html-when-copying.html:
* LayoutTests/platform/ios/TestExpectations:

Canonical link: https://commits.webkit.org/278095@main